Fix "go to label" in bazel LSP
When executing "go to definition" on a label, it would previously go to the correct file but not to the definition of that target. The reason was because we were searching for a function call where the name is the whole label. Now it parses the label to extract the name and searches using that.